Add icons for each tab. Change file to tsx to support jsx components.
